fastapi-app/app/requirements.txt
fastapi
pydantic>=2.0.0,<3.0.0
uvicorn[standard]
asyncpg
azure-storage-blob
aiohttp
aiofiles
python-multipart

psycopg2-binary
python-dotenv
alembic
pydantic-settings



python-jose[cryptography]
passlib[bcrypt]
pytz
python-dateutil
azure-cognitiveservices-speech==1.34.0
pygame
openai
tiktoken
numpy
soundfile 


pydub

